requirements.txt
pySerial
drawnow
matplotlib
numpy
